+++ include/ssog.hpp
@@ -3,19 +3,20 @@
 
 #include <vector>
 #include <Eigen/Dense>
+#include <numeric>
 
 class SparseSynthesizer {
 public:
     SparseSynthesizer(int d_in, int d_out, int n_experts, int k_active) 
         : d_in(d_in), d_out(d_out), n_experts(n_experts), k_active(k_active) {
         for (int i = 0; i < n_experts; ++i) {
-            experts.push_back(Eigen::MatrixXd::Random(d_out, d_in));
+            experts.push_back(Eigen::MatrixXf::Random(d_out, d_in));
         }
-        W_gate = Eigen::MatrixXd::Random(n_experts, d_in);
+        W_gate = Eigen::MatrixXf::Random(n_experts, d_in);
     }
 
-    Eigen::VectorXd forward(const Eigen::VectorXd& h) {
-        Eigen::VectorXd gate_logits = W_gate * h;
+    Eigen::VectorXf forward(const Eigen::VectorXf& h) {
+        Eigen::VectorXf gate_logits = W_gate * h;
         
         // Simple top-k routing
         std::vector<int> indices(n_experts);
@@ -23,10 +24,10 @@ public:
         std::partial_sort(indices.begin(), indices.begin() + k_active, indices.end(),
                           [&](int i, int j) { return gate_logits(i) > gate_logits(j); });
 
-        Eigen::VectorXd output = Eigen::VectorXd::Zero(d_out);
+        Eigen::VectorXf output = Eigen::VectorXf::Zero(d_out);
         for (int i = 0; i < k_active; ++i) {
             int idx = indices[i];
-            double weight = std::exp(gate_logits(idx)); // Simplified softmax
+            float weight = std::exp(gate_logits(idx)); // Simplified softmax
             output += weight * (experts[idx] * h);
         }
         return output;
@@ -34,16 +35,16 @@ public:
 
     void save(std::ostream& os) const {
         for (const auto& expert : experts) {
-            os.write((char*)expert.data(), expert.size() * sizeof(double));
+            os.write((char*)expert.data(), expert.size() * sizeof(float));
         }
-        os.write((char*)W_gate.data(), W_gate.size() * sizeof(double));
+        os.write((char*)W_gate.data(), W_gate.size() * sizeof(float));
     }
 
     void load(std::istream& is) {
         for (auto& expert : experts) {
-            is.read((char*)expert.data(), expert.size() * sizeof(double));
+            is.read((char*)expert.data(), expert.size() * sizeof(float));
         }
-        is.read((char*)W_gate.data(), W_gate.size() * sizeof(double));
+        is.read((char*)W_gate.data(), W_gate.size() * sizeof(float));
     }
 
 private:
@@ -51,8 +52,9 @@ private:
     int d_out;
     int n_experts;
     int k_active;
-    std::vector<Eigen::MatrixXd> experts;
-    Eigen::MatrixXd W_gate;
+    std::vector<Eigen::MatrixXf> experts;
+    Eigen::MatrixXf W_gate;
 };
 
 #endif // SSOG_HPP
+
